+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 - New features:
   - Click anywhere on toolbar to open chapter selection view and move to current book (#206)
   - Add a seek bar to set font size scale (#206)
+  - Long tap on search results to show preview
 - Changes:
   - Make view pager for reading less sensitive to swipe
 - Update dependencies: